feat: AddJumbled() for simple file/dir composing.
AddJumbled() makes the user of the library write less code to do a
really common task of adding a collection of optional files and
directories located on the same filesystem.

This also adds the multipleOptionsOption which allows returning multiple
options that look and act like a single action.  This isn't exported yet
because it's unclear if this is something that is actually needed
externally or will just be confusing.