feat!:remove key swizzling and ignoring case by default
BREAKING CHANGE: Removes AlterKeyCase() and changes the default behavior
for MatchName() to be exact match vs. ignoring character case.

Why change to be more strict?  There are a number of complexities that
arise from trying to merge keys of a structure.  For example, one key
could define processing instructions and a conflicting key could have no
processing instructions.  In this case it is unclear what goschtalt
should do.  Hence, it is better to remove the uncertainty in favor of
strictness.