feat!:Refactor unmarshal/value code to be simpler
BREAKING CHANGE: Removes the following options:

- DecodeHook()
- ErrorUnused()
- ErrorUnset(),
- WeaklyTypedInput()
- IgnoreUntaggedFields()
- ZeroFields()
- Exactly()

Prior to this change mapstructure was the expected transformation
interface for the unmarshal and value based interfaces.  This change
moves that implemention detail out of the goschtalt interface and makes
the ease of writing data converters better.

Additionally, the ability to convert objects into the configuration via
AddValue() was quite limited.  With this change it makes the conversions
work the way you would expect.