fix: Correctly reference absolute files.
This fixes issue #76 where files outside the local working directory
tree are not accessable because they either have leading / or volume
names in the way and fs.FS can't handle them.

This also adds two tests based on the project directory structure to
test that this works end to end.